New input
Set up new input with the C# class generation instead of the player input component.

With the new input class I set up multiple fire options. Click once to fire one held object, or press and hold to fire all.

Moved the playground scene so it is more visible in the scene folder.